refactor(types): move module-specific types into modules
