+++ chandragen/formatters/multiline_formatters.py
@@ -1,8 +1,8 @@
 import textwrap
 
-from chandragen.types import ConverterConfig as Config
-from chandragen.types import FormatterFlags as Flags
-from chandragen.types import MultilineFormatter
+from chandragen.formatters.types import FormatterConfig as Config
+from chandragen.formatters.types import FormatterFlags as Flags
+from chandragen.formatters.types import MultilineFormatter
 
 
 class FormatTablesAsUnicode(MultilineFormatter):
